w40k: Add 'ArmyUnit' forms
Add ``ÀrmyModelForm`` which takes a additional ``Unit`` instance in
order to limit the relevant ``model`` queryset to only those models that
a viable option for the passed unit.

Add a ``ArmyUnitForm`` which changes the default ``unit`` widget to
``hidden`` we do not want this to be user editable but rather passed
along by the get request (at least for now).

Ref.: #34