w40k: Add min/max_amount to 'UnitModel'
This commits starts a major shift in how we think of ``UnitModel``s.
Previously, the idea was that each model instance represented one
particular model for one specific final ``ArmyModel``. This seemed
very redundant and error prone, especially if a ``Unit`` contains many
of de facto identical ``UnitModel``s.
Our new approach allows us us specify a min/max amount of the given
``UnitModel`` aka model-configuration. This can be used by during
``ArmyModel`` validation. The rough idea is that the actual
``ArmyModel`` specifies which 'template' (``UnitModel``) is is 'using'.
Validation can then check not only if it fits to this template but also
if the appropiate amounts per unit add up.

Ref.: #34